stdint

+++ platform.h
@@ -1,7 +1,6 @@
 	#ifdef __unix__
-	#include <stddef.h>
 	#include <unistd.h>
-	#define STEPSIZE ((size_t)getpagesize())
+	#define STEPSIZE (getpagesize())
 	#else
 	#define STEPSIZE 4096
 	#endif
